# Writerside/api/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/admin/tape:
    get:
      tags:
      - admin-tape-controller
      description: Getting tape
      operationId: getTape
      parameters:
      - name: tapeId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found tape
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TapeResponse'
        "404":
          description: Tape could not be found.
          content:
            application/json:
              schema:
                type: string
    put:
      tags:
      - admin-tape-controller
      description: Editing tape
      operationId: editTape
      parameters:
      - name: tapeId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TapeCreationRequest'
        required: true
      responses:
        "200":
          description: Edited tape
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Tape could not be found.
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Tape already exists.
          content:
            application/json:
              schema:
                type: string
    post:
      tags:
      - admin-tape-controller
      description: Create new tape
      operationId: createTape
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TapeCreationRequest'
        required: true
      responses:
        "201":
          description: Created new tape
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Tape already exists.
          content:
            application/json:
              schema:
                type: string
    delete:
      tags:
      - admin-tape-controller
      description: Deleting tape
      operationId: deleteTape
      parameters:
      - name: tapeId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Deleted tape
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TapeResponse'
        "404":
          description: Tape could not be found.
          content:
            application/json:
              schema:
                type: string
...
  /api/admin/admin:
    get:
      tags:
      - admin-admin-controller
      description: Gets Admin
      operationId: getAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Found Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
    put:
      tags:
      - admin-admin-controller
      description: Edit Admin
      operationId: editAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSignupRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
    post:
      tags:
      - admin-admin-controller
      description: Signs up new Admin
      operationId: createAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSignupRequest'
        required: true
      responses:
        "201":
          description: Signed up new Admin
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - admin-admin-controller
      description: Deletes Admin
      operationId: deleteAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Deleted Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
...
  /api/admin/tapes:
    get:
      tags:
      - admin-tape-controller
      description: Getting all tapes from year
      operationId: getAllTapes
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Found all tapes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TapeResponse'
    delete:
      tags:
      - admin-tape-controller
      description: Deleting list of tapes
      operationId: deleteTapes
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
        required: true
      responses:
        "200":
          description: Deleted tapes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TapeResponse'
        "404":
          description: Tape could not be found.
          content:
            application/json:
              schema:
                type: string
...
components:
  schemas:
...
    TapeCreationRequest:
      type: object
      properties:
        name:
          type: string
        year:
          type: integer
          format: int32
        lk:
          type: boolean
...
    AdminSignupRequest:
      type: object
      properties:
        firstname:
          type: string
        surname:
          type: string
...
    LessonResponse:
      type: object
      properties:
        lessonId:
          type: integer
          format: int64
        day:
          type: integer
          format: int32
        hour:
          type: integer
          format: int32
    TapeResponse:
      type: object
      properties:
        tapeId:
          type: integer
          format: int64
        name:
          type: string
        year:
          type: integer
          format: int32
        releaseYear:
          type: integer
          format: int32
        lessonResponses:
          type: array
          items:
            $ref: '#/components/schemas/LessonResponse'
        lk:
          type: boolean
...
    AdminResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        adminId:
          type: integer
          format: int64
...